ci: fix remaining Rust 1.98 lints

+++ crates/rc-cli/src/resource_scope.rs
@@ -246,7 +246,7 @@ fn apply_fallback(limits: Limits) {
                 },
             ) == 0
             {
-                Some(soft as u64)
+                Some(soft)
             } else {
                 None
             }
